requirements.txt
Django>=5.1.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
Pillow>=10.0.0
django-cors-headers>=4.3.0
